% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect2Db.R
\name{connect2Db}
\alias{connect2Db}
\title{Connect to Database}
\usage{
connect2Db(db.info.ls, drv = odbc::odbc())
}
\arguments{
\item{db.info.ls}{A named list containing database connection parameters.
Must include: driver, server, database, username, password}

\item{drv}{Database driver object. Default is odbc::odbc()}
}
\value{
A database connection object
}
\description{
Establishes a connection to a database using ODBC driver with provided
connection parameters.
}
\examples{
\dontrun{
# Create connection info list
db.config.ls <- list(
  driver = "SQL Server",
  server = "localhost",
  database = "mydb",
  username = "user",
  password = "pass"
)

# Connect using default ODBC driver
con <- connect2Db(db.info.ls = db.config.ls)

# Connect using specific driver
con <- connect2Db(
  db.info.ls = db.config.ls,
  drv = RPostgreSQL::PostgreSQL()
)
}
}
