% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAcledData.R
\name{getAcledData}
\alias{getAcledData}
\title{getAcledData: Retrieve ACLED data from the API: https://api.acleddata.com/}
\usage{
getAcledData(
  api.email,
  api.key,
  start.date = NULL,
  end.date = NULL,
  n.months = NULL
)
}
\arguments{
\item{api.email}{character. ACLED user email account}

\item{api.key}{character: ACLED API key}

\item{start.date}{date: Desired start-date of data. Default: NULL}

\item{end.date}{date: Desired end-date of ACLED data. Default: NULL}

\item{n.months}{numeric: Number of months of acled data needed. Default: 24}
}
\value{
A data.frame with ACLED data from the API
}
\description{
getAcledData: Retrieve ACLED data from the API: https://api.acleddata.com/
}
