% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataTimeSpan.R
\name{getDataTimeSpan}
\alias{getDataTimeSpan}
\title{getDataTimeSpan: Get time span (start-date, end-date) for each variable 
with at least one non-missing value.}
\usage{
getDataTimeSpan(df, time.var = "YEAR", vars = NULL)
}
\arguments{
\item{df}{data.frame}

\item{time.var}{character: Name of the "time" variable in the dataset (e.g. year)}

\item{vars}{character: Name(s) of variables for which time span is needed. If 
left empty, function returns time spans of all variables in the dataset}
}
\value{
a data.frame with 3 columns: variable name, start-date, end-date
}
\description{
getDataTimeSpan: Get time span (start-date, end-date) for each variable 
with at least one non-missing value.
}
